@@ -2,7 +2,7 @@ package org.meetsl.snetbus_master
 
 import android.os.Bundle
 import android.os.PersistableBundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import org.meetsl.snetbus.NetBus
 
 /**
