@@ -2,7 +2,7 @@ package org.meetsl.snetbus_master
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import android.view.View
 import org.meetsl.snetbus.NetBus
