@@ -1,7 +1,7 @@
 package org.meetsl.snetbus_master.lifecycle.test
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import org.meetsl.snetbus.NetBus
 import org.meetsl.snetbus.NetMode
