@@ -1,6 +1,6 @@
 package org.meetsl.snetbus.lifecycle
 
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 
 /**
  * Created by meetsl
