@@ -1,8 +1,8 @@
 package org.meetsl.snetbus.lifecycle
 
 import android.content.Context
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentActivity
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
 
 /**
  * Created by meetsl
