@@ -1,9 +1,9 @@
 package org.meetsl.snetbus_master
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.util.Log
 import android.view.View
+import androidx.fragment.app.Fragment
 import org.meetsl.snetbus.NetMode
 import org.meetsl.snetbus.NetSubscribe
 
